Tidying up, norminetted except for print_contents
