+++ src/fill_object_getters.c
@@ -6,7 +6,7 @@
 /*   By: plouda <plouda@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/17 14:03:53 by plouda            #+#    #+#             */
-/*   Updated: 2023/11/20 13:33:52 by plouda           ###   ########.fr       */
+/*   Updated: 2023/11/20 15:53:21 by plouda           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
