+++ src/format_check_utils.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   format_check_utils.c                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: plouda <marvin@42.fr>                      +#+  +:+       +#+        */
+/*   By: plouda <plouda@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/13 09:13:21 by plouda            #+#    #+#             */
-/*   Updated: 2023/11/17 13:22:56 by plouda           ###   ########.fr       */
+/*   Updated: 2023/11/20 16:05:31 by plouda           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,24 +16,24 @@ static int	iter_float_triad(char *str, int *i, int *j, int flag)
 {
 	if (str[*i] == '-')
 		(*i)++;
-	if (!ft_isdigit(str[*i])) // is there at least one digit before the decimal
+	if (!ft_isdigit(str[*i]))
 		flag++;
 	while (str[*i] && ft_isdigit(str[*i]))
 		(*i)++;
 	if (str[*i] == '.')
 		(*i)++;
-	if (!ft_isdigit(str[*i]) && flag) // if no digits on either side, invalid
+	if (!ft_isdigit(str[*i]) && flag)
 		return (0);
 	while (str[*i] && ft_isdigit(str[*i]))
 		(*i)++;
 	if ((*j) != 2)
 	{
-		if (str[(*i)++] != ',') // is there a comma at the expected place
+		if (str[(*i)++] != ',')
 			return (0);
 	}
 	else
 	{
-		if (str[*i]) // is there anything after the last digit
+		if (str[*i])
 			return (0);
 	}
 	(*j)++;
@@ -43,19 +43,19 @@ static int	iter_float_triad(char *str, int *i, int *j, int flag)
 static int	iter_int_triad(char *str, int *i, int *j)
 {
 	if (str[*i] == '-')
-			(*i)++;
-	if (!ft_isdigit(str[*i])) // is there at least one digit
+		(*i)++;
+	if (!ft_isdigit(str[*i]))
 		return (0);
 	while (str[*i] && ft_isdigit(str[*i]))
 		(*i)++;
 	if (*j != 2)
 	{
-		if (str[(*i)++] != ',') // is there a comma at the expected place
+		if (str[(*i)++] != ',')
 			return (0);
 	}
 	else
 	{
-		if (str[*i]) // is there anything after the last digit
+		if (str[*i])
 			return (0);
 	}
 	(*j)++;
@@ -76,7 +76,7 @@ int	is_float_triad(char *str)
 		if (!iter_float_triad(str, &i, &j, flag))
 			return (0);
 	}
-	if (!str[i] && j < 3) // too short
+	if (!str[i] && j < 3)
 		return (0);
 	return (1);
 }
@@ -93,84 +93,8 @@ int	is_rgb_format(char *str)
 		if (!iter_int_triad(str, &i, &j))
 			return (0);
 	}
-	if (!str[i] && j < 3) // too short
+	if (!str[i] && j < 3)
 		return (0);
-	return (1);
-}
-
-int	is_floatable(char *str)
-{
-	int	i;
-
-	i = 0;
-	if (str[i] == '-')
-		i++;
-	while (str[i] && ft_isdigit(str[i]))
-		i++;
-	if (str[i] == '.')
-		i++;
-	while (str[i] && ft_isdigit(str[i]))
-		i++;
-	if (str[i])
-		return (0);
-	return (1);
-}
-
-int	is_integer(char *str)
-{
-	int	i;
-
-	i = 0;
-	if (str[i] == '-')
-		i++;
-	while (str[i] && ft_isdigit(str[i]))
-		i++;
-	if (str[i])
-		return (0);
-	return (1);
-}
-
-int	float_in_range(char *str)
-{
-	int	i;
-	int	len;
-
-	i = 0;
-	len = 0;
-	if (str[i] == '-')
-		i++;
-	while (str[i] && ft_isdigit(str[i]))
-	{
-		i++;
-		len++;
-	}
-	if (len > 4)
-		return (0);
-	len = 0;
-	if (str[i] == '.')
-		i++;
-	while (str[i] && ft_isdigit(str[i]))
-	{
-		i++;
-		len++;
-	}
-	if (len > 5)
-		return (0);
-	return (1);
-}
-
-int	triad_in_range(char *triad)
-{
-	char	**subsplit;
-
-	subsplit = ft_split(triad, ',');
-	if (!float_in_range(subsplit[0]) || !float_in_range(subsplit[1])
-		|| !float_in_range(subsplit[2]))
-	{
-		ft_free_split(&subsplit);
-		return (0);
-	}
-	ft_free_split(&subsplit);
 	return (1);
 }
