+++ src/print_contents.c
@@ -6,7 +6,7 @@
 /*   By: plouda <plouda@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/20 12:02:07 by plouda            #+#    #+#             */
-/*   Updated: 2023/11/20 12:02:42 by plouda           ###   ########.fr       */
+/*   Updated: 2023/11/20 16:06:04 by plouda           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
